move get_activities_test to testing package
